+++ triple_arena/src/chain/chain_no_gen_iterators.rs
@@ -228,24 +228,24 @@ impl<P: Ptr, T, B: ArenaBacking> ChainNoGenArena<P, T, B> {
 
     /// Iteration over all valid `P`s in the arena
     pub fn ptrs(&self) -> Ptrs<'_, P, LinkNoGen<P, T>, B> {
-        self.a.ptrs()
+        self.a.old_ptrs()
     }
 
     /// Iteration over `&LinkNoGen<P, T>`
     pub fn vals(&self) -> Vals<'_, P, LinkNoGen<P, T>, B> {
-        self.a.vals()
+        self.a.old_vals()
     }
 
     /// Mutable iteration over `LinkNoGen<P, &mut T>`
     pub fn vals_mut(&mut self) -> ValsLinkMut<'_, P, T, B> {
         ValsLinkMut {
-            iter_mut: self.a.vals_mut(),
+            iter_mut: self.a.old_vals_mut(),
         }
     }
 
     /// Iteration over `(P, &LinkNoGen<P, T>)` tuples
     pub fn iter(&self) -> Iter<'_, P, LinkNoGen<P, T>, B> {
-        self.a.iter()
+        IntoIterator::into_iter(&self.a)
     }
 
     /// Iteration over `(P, &LinkNoGen<P, T>)` tuples corresponding to all
@@ -259,18 +259,22 @@ impl<P: Ptr, T, B: ArenaBacking> ChainNoGenArena<P, T, B> {
     /// Mutable iteration over `(P, LinkNoGen<P, &mut T>)` tuples
     pub fn iter_mut(&mut self) -> IterLinkMut<'_, P, T, B> {
         IterLinkMut {
-            iter_mut: self.a.iter_mut(),
+            iter_mut: IntoIterator::into_iter(&mut self.a),
         }
     }
 
     /// Same as [crate::Arena::drain]
     pub fn drain(&mut self) -> Drain<'_, P, LinkNoGen<P, T>, B> {
-        self.a.drain()
+        let adv = self.a.advancer();
+        Drain {
+            arena: &mut self.a,
+            adv,
+        }
     }
 
     /// Same as [crate::Arena::capacity_drain]
     pub fn capacity_drain(self) -> CapacityDrain<P, LinkNoGen<P, T>, B> {
-        self.a.capacity_drain()
+        IntoIterator::into_iter(self.a)
     }
 
     /// Performs [ChainNoGenArena::compress_and_shrink] and returns an `Arena<P,