+++ triple_arena/CHANGELOG.md
@@ -18,6 +18,7 @@
 - Removed `compress_and_shrink*` in favor of just `compress*` functions, `self.reallocate_min_capacity(self.len())` can be used afterwards to retain the original functionality
 - Removed the `remove_by`, `capacity_drain`, and `*_shrink` functions. With some backings, shrinking capacity is not actually possible. The `reallocate_min_capacity` method should be used in tandem with other functions instead.
 - Added `set_generation`, `inc_generation`, `find_first_ptr`, `find_last_inx_ptr`, and other low level manipulation functions
+- Removed `FromIterator` impls because they were awkward and hide various kinds of fallibility
 
 ## [0.14.0] - 2025-02-10
 ### Changes